requirements.txt
flask
pixivpy==3.5.10
gunicorn==20.0.4
Flask-Migrate==2.5.3
